Исправление замечаний. Добавление Allure отчета

+++ src/main/java/entity/Sale.java
@@ -10,11 +10,11 @@ import java.math.BigDecimal;
 public class Sale {
 
     private Long saleId;
-    private Integer productId;
+    private Long productId;
     private Integer quantity;
     private BigDecimal price;
 
-    public Sale(Integer productId, Integer quantity, BigDecimal price) {
+    public Sale(Long productId, Integer quantity, BigDecimal price) {
         this.productId = productId;
         this.quantity = quantity;
         this.price = price;
